+++ rm/formatters/__init__.py
@@ -1,0 +1,3 @@
+from rm.formatters.stepwise import format_instruction, format_trajectory_stepwise, get_last_k_step_char_start
+
+__all__ = ["format_instruction", "format_trajectory_stepwise", "get_last_k_step_char_start"]