+++ rm/vlm/__init__.py
@@ -1,0 +1,3 @@
+from rm.vlm.wrapper import MinimalVLMRewardModel
+
+__all__ = ["MinimalVLMRewardModel"]